requirements.txt
amqp==5.3.1
annotated-types==0.7.0
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.5.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Cython==3.0.12
Flask==2.3.2
Flask-SocketIO==5.5.1

google-ai-generativelanguage==0.6.15
google-generativeai>=0.3.2
google-api-core>=2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos>=1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
proto-plus>=1.26.1
protobuf>=5.29.4,<6.0.0

h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.2
MarkupSafe==3.0.2


numpy==1.26.4
pandas==2.0.3

pdfkit==1.0.0
prompt_toolkit==3.0.50
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
rsa==4.9
setuptools==78.0.2
simple-websocket==1.1.0
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
